Modify ndb_read_balance test to skip if the Query Cache is enabled.
Having Query cache enabled adds an extra table scan to each PK lookup.
This affects the results as they show the low level table operations
occurring for a given SQL statement.
The test is intended to verify read routing and balance, which is determined
at layers below the Query Cache.
So it seems reasonable to disable this test when the QC is enabled, to avoid
the test failure notification.