Bug #23602217:  MISSES TO USE OLDER LCP WHEN LATEST LCP IS NOT RECOVERABLE
Problem:
--------

When a fragment is to be restored it can mostly use LCPs. In 7.4 there are
upto 2 LCPs that can be recoverable. However sometimes an LCP is reported
as completed to DIH and then crashes before the last GCI written into this
LCP has completed. In this case one of the LCPs is not restorable and the
older LCP should be used, which was not happening thus resulting in longer
restart times.

Fix:
----

Ensured that the older LCP is picked up to be restored. Also added test
coverage to check that the older LCP was picked up and there were no bugs
in the new flow of execution.