Fix warnings from innobase_build_index_translation(), which say there is a column info mismatch.
The problem is that there are two mappings, one is
get_innobase_type_from_mysql_type() and the other is in create_table_metadata(
before this patch). So some mtypes are mismatched.

This patch can fix this issue, however we have to revisit this problem later.
Because current fix in create_table_metadata() duplicate some existing code.
Had better to make them clean.