Bug#22202516: ENABLING SEMI-SYNC DURING COMMIT CAN CAUSE MASTER TO ASSERT.
Problem:
========
To be more specific, the problematic code is in
ReplSemiSyncMaster::commitTrx() , line 792 , with
assert(entry);

In fact, 'entry' is not NULL only if semi-sync was enabled
when the commit phase started. If semi-sync is enabled
between commit and after_commit() then 'entry' will be NULL
and the above assert will cause mysqld to assert.

Analysis:
=========
In semi sync replication a transaction that gets committed
on master will wait for an acknowledgement from slave.
In order to achieve this there are few steps that are
carried out during different stages of commit. First comes
flush stage. One 'Entry' node (active transaction node) is
created and it is associated with a single transaction or a
group of transactions that got flushed at the flush stage.
This object gets inserted into active transaction list. Next
comes sync stage. After sync stage the commit is finalized.
At this stage for each thread trx_wait_binlog_name and
trx_wait_binlog_pos are set with respect to their
corresponding binary log positions. While the dump thread
sends these transactions across the network and receives
their ack,  the thread that performed commit has to wait. In
order to do that the committed thread gets its
trx_wait_binlog_name and trx_wait_binlog_pos and tries to
retrieve the corresponding entry node that is associated
with this position. There are two scenarios where the Entry
node may not be present.

case 1: Where the transaction was already acknowledged as
the slave's reply positions are higher than the current
threads waiting positions.

case 2: When transaction entered into flush stage of
ordered_commit semi sync was not enabled and there was no
corresponding active transaction node that got created.

Case 1 will never get to waiting stage as it has already
got the ack. It will simply go away. But case 2 will result
in the above assert as it is trying to wait for an ack.

Transactions which fall under case 2 should not wait for
ack. They should simply exit and they should be considered
as async transactions.

Fix:
===
If a transaction doesn't find a corresponding 'Entry' in the
active transaction list and it has not received
acknowledgement from slave then this transaction will not be
considered as semi sync transaction and it will not wait for
an ack.