Bug#24337411: MYSQL DEFAULTS-EXTRA-FILE AND -LOGIN-PATH OPTIONS               NOT WORKING WITH MTR --RESTART
Issue:
------
Passing the --defaults-extra-file or the --login-path option
in a MTR test to restart the server results in an "unknown variable"
error because option file options need to be passed to the server
before any other variable or option.

Fix:
----
A new sub routine which processes these option file options
before the mysqld variables has been added to the perl script.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
Reviewed-by: Lalit Choudhary <lalit.choudhary@oracle.com>
RB: 13535