Bug #24453571 	SERVER CRASHES WHEN INSTALL COMPONENT IS ISSUED,                AFTER MYSQL_UPGRADE
Post push changes.
component.test is failing because of the component_id and
component_group_id are different in result file compared to the
test run.
When we run the tests component-upgrade.test and component.test
together, mtr won't restart the server for the second test. Hence the
component_id in the 'select' statement for component.test is showing
different than in the value in result file.
As fix, the 'select *' is updated with 'select count(*)'