WL#6599 - New Data Dictionary and I_S integration.
This is a post-push patch to fix memory leak identified by the
ASAN in the unit test.

Issue here is, memory allocated for Dictionary_impl class's
instance is not deallocated in the unit test.
Modified code to deallocate memory now to fix this issue.
Even modified code to create instance of Dictionary_impl class
only once at the test start and deallocate at the end.