Bug#23106330: MISSING CHECK_STACK_OVERRUN IN JSON_BINARY::SERIALIZE_JSON_VALUE
The JSON tests ran out of stack space on some platforms when running
against a debug enabled server. Because of lacking checks for stack
usage, the server would not detect this situation in time, and it led
to the server exiting.

The fix is twofold:

1) It add calls to check_stack_overrun() before attempting to
serialize a nested array or object in order to detect stack overruns
up front and fail gracefully when processing deeply nested JSON
documents. This prevents the server from exiting if it runs out of
stack space.

2) It reorganizes the code that serializes JSON document so that it is
less stack-hungry when compiled without optimization. Specifically:

  a) The function json_binary::append_value() is manually inlined the
  two places it is called. This function is called recursively when
  serializing nested JSON documents, so inlining it removes one stack
  frame for each level of nesting in the document.

  b) The code in json_binary::serialize_json_object() that writes the
  key entries section of the object, is moved out to a separate
  function called append_key_entries(). This reduces the size of the
  stack frames for serialize_json_object(), which is called
  recursively when serializing nested objects.

  c) The cases for serializing opaque, datetime and decimal values in
  json_binary::serialize_json_value() are moved out to separate
  functions. This reduces the stack frame size of the recursive
  serialize_json_value() function.

This makes the JSON tests pass with the default stack size on more
platforms.