BUG#23582603 - SOCKET LOCKFILE IS NOT CLEANED UP IF WRITE FAILS                 , LEADING TO CRASHLOOPING.
During server startup, the pid write to the lock file
mysql.sock.lock fails (due to errros like disk full).
The server doesn't delete it and return error. This
may cause next server restart to fail unless the file
is manually cleaned up.

This fix shall be to try and unlink the socket lock file on
failure of write, fsync and close failure.