Bug #23615208 ASSERT IN INNODB: ASSERTION FAILURE: UT0UT.CC:702
Problem:

This is a crash recovery scenario.  The core dump shows that the
LOB reference object is all zeroes.  So it could be an insert or
an update operation that has crashed in between and hence the LOB
reference is null.  During recovery, there is an attempt to fetch
the LOB pages using this null LOB reference.  Hence this assert.

Solution:

During a fetch operation, check if the LOB reference is null.

rb#13675 approved by Satya.