This is a patch for the bug#24449174 -- dropping a partition drops the trigger on 8.0.0.
In case there is a trigger assigned to a partitioned table, attempt to
drop the partition led to dropping any trigger associated with the table.

We solve this problem by moving triggers from old table definition to
a new one, before replacing former with the latter.

All places where dd::rename_table<dd::Table> is called were checked.
While doing it a dead block was detected that calls dd::rename_table.
The block is located in the function mysql_update_dd() and it was removed.
Additionally, the function sync_ddl_log() was dropped since it isn't called
anymore.