Fix the issue that during ALTER TABLE, the order of indexes could be different because one unique index is added, or an index name is renamed.
find_index(): New function, to find the proper InnoDB dict_index_t
object according to given dd::Index. The first one of both are
always the PRIMARY KEY, but the name may not be exact the index
name due to it's the first unique index upgraded to PRIMARY KEY.
This function also check the renaming index name case.