WL#6599 - New Data Dictionary and I_S integration.
This is a post-push patch to fix test failures with NDB on
embedded server.

The changes made to ignore 'ndbinfo' and its tables when ndb
cluster is not running or system variable "show_hidden"(exported
by NDB)is not set, are compiled out for the embedded server.
Hence on embedded server with NDB plugin, schema 'ndbinfo'
and its tables are listed always. This caused extra test
failures on embedded server with NDB.

To fix this issue, including the changes made to ignore
'ndbinfo' schema and test tables for embedded server too.