Bug #24488219: INCLUDE WHAT YOU USE
First small patch out of a series of many, restricted to strings/
only to start out small.

The general gist of these is to install the include-what-you-use
tool (e.g. by “sudo apt install iwyu” in Debian), then cmake with

  -DCMAKE_C_COMPILER=/home/sgunders/iwyu-gcc -DCMAKE_CXX_COMPILER=/home/sgunders/iwyu-g++

iwyu-gcc is a program that runs IWYU, outputs (atomically) its output
to stdout, and then actually compiles the file using Clang to make sure
that compilation can proceed. It looks like this:

  #! /bin/sh
  T=$(tempfile)
  /usr/bin/iwyu "$@" > $T
  lockfile /tmp/iwyu.lock
  cat $T
  rm -f /tmp/iwyu.lock $T
  exec /usr/bin/clang "$@"

and analogously for g++, just with clang replaced by clang++.

The stdout output is stored in a log file and given to fix_include:

  /usr/bin/fix_include --ignore_re='(extra|zlib|libedit|innobase|obj|libevent|regex|myisam|rapid/plugin/x)/' < log

After that, the patch is manually inspected and corrected where needed.
Many changes work out-of-the-box, but some patterns tend to confuse
the tool, so it sometimes needs manual correction.

Change-Id: Ie194ac838c7d3b615ebb2e8b5fea68cce4b0e213