Bug #24480254 LOB DELETER NOT DELETING THE LOB PAGES
Problem:

While doing code reading I identified the following problem:
Code in lob::Deleter::destroy() function is as follows:

1708
1709                 while (m_ctx.m_blobref.page_no() == FIL_NULL) {
1710                         err = free_first_page();
1711                         if (err != DB_SUCCESS) {
1712                                 break;
1713                         }
1714                 }
1715

The while condition should contain != instead of ==.  This means that
the LOB deleter is not really deleting the LOB pages.  This is a
regression introduced by me in WL#9141 InnoDB: Refactor uncompressed
BLOB code to facilitate partial fetch/update.

Solution:

The while condition should check that the page_no in the LOB
reference is not equal to FIL_NULL.

WIP.