Bug#23297190 : RPL_GTID_SERVER_SIGHUP AND RPL_MASTER_POS_WAIT_AFTER_STARTUP FAILS IN VALGRIND
In the testcases we are restarting the master server which is taking more time
on slow platforms like valgrind and slave fails to reconnect with
"error reconnecting to master '$Master_Port'- retry-time: 1 retries:10".

Fix :

Changed master_connect_retry=30 and master-retry-count=30 on slave. Now the
slave retries connecting after every 30 secs for 30 times which gives
sufficient time to reconnect to the master.