Bug#23251423 DBSPJ::LOOKUP_ABORT() DOES NOT REMOVE ANY 'M_DEFERRED' OPERATIONS.
In order to avoid that new lookup REQ's are produced
by SPJ for an aborted Request, the 'm_deferred' list
of operations to be resumed later should be cleared when aborted.

Failing of doing so may result in aborted lookup REQ's still
sending row requests to the datanodes, which is wasted work.
There are no (known) failures as such caused by this defect.