Bug#23251145 DBSPJ::EXECTRANSID_AI() FAILURE WHEN SPJ REQUEST BEING ABORTED
Patch 2/2

The error/ abort handling in Dbspj::execTRANSID_AI() was implemented
such that ::abort() was called prior to have completed processing
of the incomming signal. As ::abort() sends signals to the LDM, it
partly overwrote the contents of the signal which was later
required by ::execTRANSID_AI().

The root cause was parly related to a rather add-hock pattern
used for detecting completion and/or abort of REQs sent
from the SPJ block to LDM /DIH: Generally such a request
has to wait for a number of outstanding replies (CONF/REF)
before it reach its 'completion state'. When 'completed'
the SPJ block may proceed to either its next step, or
send a client reply, and clean up after the request.

NOTE: Even when a SPJ Request is aborted, we have to wait for all
outstanding replies before sending a reply back to the client.

The logical order of handling the above logic would be:

1) Lookup the TreeNode and Request object for the received signal.
2) Update number of outstanding replies / completion state of
   the SPJ-Request and TreeNode
3) Do whatever this particular Dbspj::exec<FOO> is supposed to do.
   Could possibly fails and call Dbspj::abort().
4) As it very last action, each Dbspj::exec<FOO> should check if
   the current (batch of) SPJ operations has completed and generate
   the appropriate client reply, and clean up any resources being
   obsolete.

However, this sequence of operation has been mixed together:

- ::checkPrepareComplete() and ::checkBatchComplete() which implemented
  4), also partly did the counting of outstanding requests, 2)
  (However, not always).
- ::checkPrepareComplete() and ::checkBatchComplete() was not always
  called as the very last action. Thus there were code paths where
  it could be called twice!. (Like when a request aborted)
- No clear pattern of which relative order 2), 3), 4) happened in.

Furthermore, the different kind of SPJ TreeNodes had its own way to
do the 'outstanding' counting. Which made code reuse and a common
pattern for completion handling harder to implement.

This had resulted in the code path for ::execTRANSID_AI() to be
particular unclear: Even after an abort, it had to continue into
m_execTRANSID_AI() - only in order to get the outstanding counting
correct.

In order to fix the bug in Dbspj::execTRANSID_AI(), the outstanding
counting for the different SPJ TreeNode types has been implemented
in the pseudo 'virtual' method *m_countSignal().
(lookup_countSignal, scanFrag_countSignal() & scanIndex_countSignal())
This method is then called as the first part of ::execTRANSID_AI(),
*before* a possible ::abort() may alter the signal contents.

As a side effect of this refactorication, the previous 'virtual'
method *m_execTRANSID_AI() became obsolete.

This patch further cleans up the usage of checkPrepareComplete() and
checkBatchComplete(): They should now only be called as the last part
of any ::exec<FOO> method possibly completing a Request. (Not from
other methods called from these)

The outstanding counting previously offered as an option through
these methods has been removed.

Renamed Dbspj::start() to ::prepare() which better describes what
it does, and relates it to ::checkPrepareComplete().