Bug#22991924 GCOLS: UTF16 STRING OPERATIONS GIVE DOUBLED \0\0 STRINGS ON EACH TABLE REBUILD
When a table is being rebuilt, the generated column
expressions are printed and re-parsed. The problem is, a
character set conversion may occur every time. This leads to
the string being converted over and over each time the table
is recreated. For certain conversions, such as utf8 to
utf16, this leads to an ever growing string. The behavior
depends on character_set_client.

The fix is to make the data dictionary always print an
explicit character set introducer when printing out the
generated column expression. This ensures that conversion
only happens once.

This obviously changes the results of generated columns wrt
charset conversions after table rebuilds.

Also, some regular expression in the X plugin are augmented
to handle charset introducers.