Bug#21286722 MISSING SANITY CHECKS FOR MALLOC() IN RESTORE.CPP, DBUTIL.CPP, SQLCLIENT.CPP
PART 1: RESTORE.CPP

In RESTORE.CPP when we try to create the Constructor for
RestoreDataIterator
it calls system malloc and does not check for failure. As this can lead
to crash, have added a validation function(validateRestoreDataIterator,
validateBackupFile) which checks for the malloc failure.

Hence, whenever we try to create an instance of class RestoreDataIterator
it is mandatory to call validateRestoreDataIterator and
validateBackupFile.
Currently it’s been used only once. So, have added validation check in
that location.