BUG#22364401: COM_QUERY RESPONSE METADATA WITH '*' IN FIELD ORG_TABLE
Problem:
In mysql-5.7, orig_table_name for a field in a derived table is displayed
as "*" resulting in failed queries which depend on this value.

Analysis:
In mysql-5.7, a derived table is merged if possible. This makes the
orig_table_name for the fields to be "*"(internal_table_name).
However, in mysql-5.6 a temptable was used for processing the derived
table. This made the orig_table_name to be updated correctly.

Solution:
While creating a view field, use table_name from "field" object to create
the Item_direct_view_ref object. Also internal_table_name has been changed
to an empty string now.
With this patch, "org_table" contained in the MYSQL_FIELD structure will
see the changes as mentioned below.
If a column is selected from view, org_table names the view. If column is
selected from a derived table, org_table will be the base table name.
In the case where a derived table is wrapping a view, org_table still names
the base table name.