Bug #24573230: REMOVE #DEFINES WITH GLOBAL INFLUENCE FROM MY_GLOBAL.H
my_global.h should not have #defines that modify behavior of system
headers; this is brittle, and better suited to CMake options,
so move it there.

It should also not try to #undef symbols from other files, which is
again brittle (especially as these symbols tend to change without
people remembering to update my_global.h -- it seems to already have
happened during the introduction of yaSSL). Instead, make sure these
symbols are simply not added during compilation in the first place.
Every plugin now gets compiled with ${SSL_DEFINES} added automatically,
so they don't have to add that manually -- except if they are recompiled
for embedded, in which case ${SSL_DEFINES} is never added in the first
place (instead of having to #undef them).

I've double-checked that the logic for EMBEDDED_LIBRARY is the same
by manually adding #error directives in my_global.h and compiling
(then removed them again before submission).

Change-Id: I70c0e2db82a9b65a1b0a5e35b75ce861bc1cdc0a