Bug#23735996 LCP NEVER COMPLETED AFTER CLEAN NODE SHUTDOWN
Change DICT master-takeover procedure to only consider schema transactions
in state TS_ENDING when deciding if rollforward or rollback. Transactions
in TS_COMPLETING are allowed to proceed and do the proper cleanup which
will avoid the LCP getting stuck. The rollforward starting point was missing
for single operation transactions causing error when now proceeding with
transactions in the TS_COMPLETING phase. When a table is being dropped we
must also support listing table while dropp is in progress, this includes
handling of finding index tables about to be dropped where there is no parent table.
DIH is modified to handle the case of tables being dropped during the master takeover.