Fix to make testcase more predictable.
Required as a follow up for patch for bug 24481551,

  CONDITION SIGNALING MISSING WHEN WAITING FOR EPOCH TO BE WRITTEN TO BINLOG

That patch removed delay in waiting for binlog to be written,
such that when checking server 2 binlog our insert had not had
the chance to reach the log yet.

Thus we have to add a 100ms sleep to guarante that some time passes
between insert ond checking binlog contents.