Bug#23048816 NDB: SPJ HAS SIGNAL_DROPPED_REP HANDLER WHICH IS NOT BOUND IN
Patch 1/2 containing the testcase.

The test case builds on an existing mechanism in
TransporterCallback.cpp and LongSignal.cpp where
the variable 'ErrorSignalReceive' could be set to '1'
if we want LongSignal.cpp to emulate failure to allocate
memory for long signals. This is enhanced to use
ErrorSignalReceive to which receiver block we should
emulate allocation failures for.