BUG#23518923 MYSQL_UPGRADE FAILS IN 7.5.2 WITH DUBIOUS ERROR NDBINFO ENGINE NOT LOADED
The bug is not on mysql_upgrade rather on the mysqld --initialize phase.
Using option "--ndbcluster" with "mysqld --initialize" corrupts the
ndbinfo table's frm files.

When "--ndbcluster" is used during the initialize step, it is ignored
during the ndbcluster plugin init phase as its not required to connect
to the cluster during the initialize step. But its handler state is not
properly set to SHOW_OPTION_DISABLED. As a result, the
ha_initialize_handlerton function erroneously assigns ndbinfo's db_type
value to the now disabled ndbcluster dataengine. And when ndbinfo gets
inited next, it gets a bad db_type value. This leads to the corruption
of the ndbinfo table's frm files.

This commit fixes that by properly disabling the ndbcluster plugin.