Bug #24586713: MAKE CLIENT ABI INDEPENDENT OF MEM_ROOT AND PSI
Remove MEM_ROOT and PSI, which are server-side concepts, from the
visible client ABI. In particular, MEM_ROOT elements in the MYSQL struct
are now changed to pointers and forward-declared, so that we can change
its definition without breaking the size or layout of said struct.
(Unfortunately, since it's a C struct, we can't make them a unique_ptr,
which would be nice to avoid sharing, but you can't have it all,
and we've already eliminated all by-value copies of MEM_ROOTs
in the previous patch.)

Also make sure we don't do ABI checks for plugins when building
without the server, which unbreaks builds with -DWITHOUT_SERVER=1.

Change-Id: Ia614e61518f036b4ad48d1d7ca609177d845be89