Bug #24557925: MYSQL_CONFIG_EDITOR CAN MAKE SERVER UNBOOTABLE
By default all mysys based programs are sourcing the login file.
The server should not be doing this since it's not a client program.

Thus added a new global to mysys, that, when set to FALSE disables
the sourcing of the login path file.
And then changed mysqld to set that variable to FALSE from its
mysys default TRUE.

Added a test case.