Bug#22923426 LOG_FILE_SIZE_1 FAILS TO REPORT INNODB: CANNOT RESIZE LOG FILES IN READ-ONLY
The InnoDB change buffer is deleted lazily in a background task.
If a previously completed test has caused something to remain in
the InnoDB change buffer on the server instance, restarting the
server with --innodb-read-only would be refused with an unexpected
error message:

"Change buffer must be empty when --innodb-read-only is set."

This WL#6494 test is expecting another error message, for InnoDB
refusing to resize the InnoDB redo logs in read-only mode.

The test failed to initiate a slow shutdown (forcing a change buffer merge)
before restarting the server in --innodb-read-only mode.

Reviewed-by: Vinay Fisrekar <vinay.fisrekar@oracle.com>